refactor: move RoundedEndsButton.kt to core presentation component
